fix: renames area-tile to area-card
