+++ src/components/area-card/AreaCardContainer.js
@@ -1,9 +1,9 @@
 import { connect } from 'react-redux';
-import AreaTile from './AreaTile';
+import AreaCard from './AreaCard';
 
 const mapStateToProps = ({ areas }, { id }) => {
   const area = areas.area[id] && areas.area[id].attributes;
   return { area: { ...area, id } };
 };
 
-export default connect(mapStateToProps, null)(AreaTile);
+export default connect(mapStateToProps, null)(AreaCard);
